% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.gpData.r
\name{add.gpData}
\alias{add.gpData}
\title{Join two \code{gpData} objects}
\usage{
add.gpData(gpData1, gpData2)
}
\arguments{
\item{gpData1}{A \code{gpData} object with at least elements \code{geno} and
\code{map}}

\item{gpData2}{Second \code{gpData} object with at least elements
\code{geno} and \code{map}}
}
\value{
A \code{gpData} object is returned containing \code{gpData1}
and \code{gpData2}
}
\description{
Function for joining two \code{gpData} objects
}
\details{
The function writes a vcf file. The format of the output is "GT". Other
formats are not supported.
}
\examples{

\dontrun{
add.gpData(maize, maize)
}

}
\seealso{
\code{\link{create.gpData}} \code{\link{codeGeno}}
}
\author{
Hans-Juergen Auinger
}
