% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synbreed-package.R
\docType{data}
\name{cattle}
\alias{cattle}
\title{Dairy cattle data}
\format{
object of class \code{gpData}
}
\description{
Data set contains genotypic, phenotypic, map and pedigree data of 500 bulls.
All individuals are labeled with an unique ID, starting with ID1430 and
ending with ID1929. Genotypic and pedigree data are from a real cattle
data set while phenotypes were simulated.  Pedigree information is
available for parents and grandparents of the phenotyped
individuals.
}
\details{
There are two quantitative phenotypes in this data set. The heritabilities of these
traits, 0.41 and 0.66, were estimated with a pedigree-based animal model using
the data on hand.

Genotypic data consists of 7250 biallelic SNP markers for every phenotyped
individual with missing data included. SNPs are mapped across all 29
autosomes. Distances in the SNP map are given in megabases (Mb).
}
\examples{

\dontrun{
library(synbreed)
data(cattle)
summary(cattle)
}

}
\keyword{datasets}
