\name{[.GenMap}
\alias{[.GenMap}
\docType{data}
\title{
Extract or replace part of map data.frame
}
\description{
Extract or replace part of an object of class \code{GenMap}.
}
\usage{
\method{[}{GenMap}(x,\dots)
}
\arguments{
  \item{x}{
object of class ''GenMap''
}
  \item{...}{
indices
}
}
\examples{
\dontrun{
data(maize)
head(maize$map)
}
}
