\name{f}
\Rdversion{1.1}
\alias{f}

\title{
a function to square a number
}
\description{
\deqn{f(x) = x^2} 
}
\usage{
f(x)
}

\arguments{
  \item{x}{
a real number or a vector of real numbers
}
}
\details{

}
\value{
 a real number or a vector of real numbers
}
\references{

}
\author{
Valentin Wimmer
}
\note{
no  complex numbers allowed
}




\examples{

f(2)

}

