\name{maize}
\alias{maize}
\docType{data}
\title{
Simulated maize data
}
\description{
This is a simulated dataset from breeding sheme for maize. 1250 doubled haploid (DH) lines were genotyped with 1117 ploymorphic SNP markers and phenotyped in a testcross with a single tester for one quantitative trait. Markers are spread along all 10 chromosomes of maize. Pedigree information is available up to 15 generations. The 1250 lines belong to 25 full sib families with 50 individuals in each family. For simulation 1000 QTL with equal and additive (no dominance or epistasis) effects were used.
}                                                                                                                                                                                                                 
\usage{data(maize)}
\format{
Object of class \code{gpData}
}
                                
\examples{
data(maize)
summary(maize)                                       
}                                  
\keyword{datasets}
