% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanPlot.r
\name{manhattanPlot}
\alias{manhattanPlot}
\title{Manhattan plot for SNP effects}
\usage{
manhattanPlot(
  b,
  gpData = NULL,
  colored = FALSE,
  add = FALSE,
  pch = 19,
  ylab = NULL,
  colP = NULL,
  ...
)
}
\arguments{
\item{b}{object of class \code{gpMod} with marker effects or numeric vector
of marker effects to plot}

\item{gpData}{object of class \code{gpData} with map position}

\item{colored}{\code{logical}. Color the chromosomes?. The \code{default} is
\code{FALSE} with chromosomes distinguished by grey tones.}

\item{add}{If \code{TRUE}, the plot is added to an existing plot. The
default is \code{FALSE}.}

\item{pch}{a vector of plotting characters or symbols: see
\code{\link[graphics]{points}}. The default is an open circle.}

\item{ylab}{a title for the y axis: see \code{\link[graphics]{title}}.}

\item{colP}{color for the different chromosomes and points.}

\item{\dots}{further arguments for function \code{plot}}
}
\description{
Plot of SNP effects along the chromosome, e.g. for the visualization of
marker effects generated by function \code{gpMod}.
}
\examples{

\dontrun{
library(synbreedData)
data(mice)
# plot only random noise
b <- rexp(ncol(mice$geno), 3)
manhattanPlot(b, mice)
}

}
\author{
Valentin Wimmer and Hans-Juergen Auinger
}
\keyword{hplot}
