% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synbreed-package.R
\docType{data}
\name{[.GenMap}
\alias{[.GenMap}
\title{Extract or replace part of map data.frame}
\arguments{
\item{x}{object of class ''GenMap''}

\item{...}{indices}
}
\description{
Extract or replace part of an object of class \code{GenMap}.
}
\examples{

\dontrun{
data(maize)
head(maize$map)
}

}
