% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synbreed-package.R
\docType{data}
\name{[.relationshipMatrix}
\alias{[.relationshipMatrix}
\title{Extract or replace part of relationship matrix}
\arguments{
\item{x}{object of class 'relationshipMatrix'}

\item{...}{indices}
}
\description{
Extract or replace part of an object of class \code{relationshipMatrix}.
}
\examples{

\dontrun{
data(maize)
U <- kin(codeGeno(maize), ret = "realized")
U[1:3, 1:3]
}

}
