% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cvData.r
\name{summary.cvData}
\alias{summary.cvData}
\alias{print.summary.cvData}
\title{Summary of options and results of the cross validation procedure}
\usage{
\method{summary}{cvData}(object, ...)
}
\arguments{
\item{object}{object of class "cvData"}

\item{...}{not used}
}
\description{
\code{summary} method for class "cvData"
}
\seealso{
\code{\link{crossVal}}
}
\author{
Theresa Albrecht
}
