% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gpData.r
\name{summary.gpData}
\alias{summary.gpData}
\alias{print.summary.gpData}
\title{Summary for class gpData}
\usage{
\method{summary}{gpData}(object, ...)
}
\arguments{
\item{object}{object of class \code{gpData}}

\item{...}{not used}
}
\description{
S3 \code{summary} method for objects of class \code{gpData}
}
\examples{

\dontrun{
library(synbreedData)
data(maize)
summary(maize)
}

}
\author{
Valentin Wimmer
}
