% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpMod.r
\name{summary.gpMod}
\alias{summary.gpMod}
\alias{summary.gpModList}
\alias{print.summary.gpMod}
\alias{print.summary.gpModList}
\title{Summary for class gpMod}
\usage{
\method{summary}{gpMod}(object, ...)
}
\arguments{
\item{object}{object of class \code{gpMod}}

\item{\dots}{not used}
}
\description{
S3 \code{summary} method for objects of class \code{gpMod}
}
\examples{

\dontrun{
library(synbreedData)
data(maize)
maizeC <- codeGeno(maize)
# marker-based (realized) relationship matrix
U <- kin(maizeC,ret="realized")/2

# BLUP model
mod <- gpMod(maizeC,model="BLUP",kin=U)
summary(mod)
}

}
\seealso{
\code{\link{gpMod}}
}
\keyword{methods}
