% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pedigree.r
\name{summary.pedigree}
\alias{summary.pedigree}
\alias{print.summary.pedigree}
\title{Summary of pedigree information}
\usage{
\method{summary}{pedigree}(object, ...)
}
\arguments{
\item{object}{object of class ''pedigree''}

\item{...}{not used}
}
\description{
\code{Summary} method for class ''pedigree''
}
\examples{

# plant pedigree
ped <- simul.pedigree(gener = 4, 7)
summary(ped)

# animal pedigree
ped <- simul.pedigree(gener = 4, 7, animals = TRUE)
summary(ped)
}
\author{
Valentin Wimmer
}
