% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.relationshipMatrix.r
\name{summary.relationshipMatrix}
\alias{summary.relationshipMatrix}
\alias{print.summary.relationshipMatrix}
\title{Summary of relationship matrices}
\usage{
\method{summary}{relationshipMatrix}(object, ...)
}
\arguments{
\item{object}{object of class ''relationshipMatrix'}

\item{...}{not used}
}
\description{
\code{Summary} method for class ''relationshipMatrix''
}
\examples{

\dontrun{
library(synbreedData)
data(maize)
U <- kin(codeGeno(maize), ret = "realized")
summary(U)
}

}
\author{
Valentin Wimmer
}
