% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.map.r
\name{summaryGenMap}
\alias{summaryGenMap}
\title{Summary of marker map information}
\usage{
summaryGenMap(map, cores = 1)
}
\arguments{
\item{map}{\code{data.frame} with columns \code{chr} and \code{pos} or a
\code{gpData} object with element \code{map}}

\item{cores}{\code{numeric}. Specifies the number of cores for parallel
computing.}
}
\value{
A \code{data.frame} with one row for each chromosome and the
intersection of all chromosomes and columns \item{noM}{number of markers}
\item{range}{range of positions, i.e. difference between first and last
marker} \item{avDist}{avarage distance of markers} \item{maxDist}{maximum
distance of markers} \item{minDist}{minimum distance of markers}
}
\description{
This function can be used to summarize information from a marker map in an
object of class \code{gpData}.  Return value is a \code{data.frame} with one
row for each chromosome and one row summarizing all chromosomes.
}
\details{
Summary statistics of differences are based on euclidian distances between
markers with non-missing position in \code{map}, i.e. \code{pos!=NA}.
}
\examples{

\dontrun{
library(synbreedData)
data(maize)
summaryGenMap(maize)
}

}
\seealso{
\code{\link{create.gpData}}
}
\author{
Valentin Wimmer
}
