% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synbreed-packageX.R
\docType{package}
\name{synbreed-package}
\alias{synbreed-package}
\title{synbreed package: A Framework for the Analysis of Genomic
Prediction Data Using R}
\description{

}
\details{
Main Functionality: (fraft form)

- Data Preparation
- Visualisation
- `create.gpData`
- `gp.Mod`
- Genetic Modelling
- and more
}
\examples{
# (better put something here)
}
\keyword{package}
