\name{LDMap}
\Rdversion{1.1}
\alias{LDMap}

\title{
LD Heatmap
}
\description{
Visualization of pairwise LD in LD heatmap for each linkage group (chromosome) using the \code{LDheatmap} package.
}
\usage{
LDMap(LDmat,gpData,chr=NULL,file=NULL,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{LDmat}{
Object of class \code{LDmat} generated by function \code{pairwiseLD}.
}
  \item{gpData}{
Object of class \code{gpData} that was used to infer \code{LDmat}.
}
  \item{chr}{
\code{numeric}. Return value is a plot for each chromosome in \code{chr}.
}  
  \item{file}{
Optionally a path to a file were plot is saved as \code{pdf}.
}
  \item{\dots}{
Further arguments that could be passed to function \code{LDheatmap}.
}
}

\seealso{
\code{\link{pairwiseLD}}, \code{\link[LDheatmap]{LDheatmap}}, \code{\link{LDDist}}
}
\examples{
data(maize)
maizeC <- codeGeno(maize)

# LD for chr 1
maizeLD <- pairwiseLD(maizeC,chr=1,type="matrix")
LDMap(maizeLD,maizeC)

}
\keyword{hplot}
