\name{simul.pedigree}
\alias{simul.pedigree}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Simulation of pedigree structure
}
\description{
This function could be used to simulate pedigree structure for a given number of generations. Random mating is assumed and inbreed are generated by chance.
}
\usage{
simul.pedigree(generations = 2, ids = 4)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{generations}{
integer. Number of generations to simulate
}
  \item{ids}{
integer or vector of integers. Number of genotypes in each generation. If length equal one, the same number will be replicated and used for each generation.
}
}
\details{                                
The parents for the F1 will be randomly choosen out of the genotypes in F0. If Par1 = Par2, an inbreed is generated.
}
\value{
An object of class \code{pedigree} with N=sum(ids) genotypes.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Valentin Wimmer
}
\note{
in developement 
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{simul.phenotype}}, \code{\link{create.pedigree}}, \code{\link{plot.pedigree}}
}
\examples{
p <- simul.pedigree(gener=4,ids=c(3,5,8,8))
plot(p)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{pedigree}
\keyword{simulation}
