\name{simul.pedigree}
\alias{simul.pedigree}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Simulation of pedigree structure
}
\description{
This function could be used to simulate pedigree structure for a given number of generations. Random mating is assumed and inbreed are generated by chance.
}
\usage{
simul.pedigree(generations = 2, ids = 4, animals=FALSE,familySize=1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{generations}{
integer. Number of generations to simulate
}
  \item{ids}{
integer or vector of integers. Number of genotypes in each generation. If length equal one, the same number will be replicated and used for each generation.
}
  \item{animals}{
logical. Should a pedigree for animals be simulated? See details. 
}  
  \item{familySize}{
numeric. Number of individuals in each full-sib family in the last generation.
}  
}
\details{                                
If \code{animals=FALSE} the parents for the F1 will be randomly choosen out of the genotypes in F0. If Par1 = Par2, an inbreed is generated. If \code{animal=TRUE} each  ID is either sire or dam. Each ID is progeny of a sire and a dam. Therefor a variable containing the sex is returned.
}
\value{
An object of class \code{pedigree} with N=sum(ids) genotypes.
}
\author{
Valentin Wimmer
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{simul.phenotype}}, \code{\link{create.pedigree}}, \code{\link{plot.pedigree}}
}
\examples{
# example for plants
ped <- simul.pedigree(gener=4,ids=c(3,5,8,8))
plot(ped)
#example for animals
peda <- simul.pedigree(gener=4,ids=c(3,5,8,8),animals=TRUE)
plot(peda)
}
