\name{summary.cvData}
\alias{summary.cvData}
\alias{print.summary.cvData}
\title{
Summary of options and results of the cross validation procedure
}
\description{
\code{summary} method for class "cvData"
}
\usage{
\method{summary}{cvData}(object,...)
}

\arguments{
\item{object}{object of class "cvData"}
\item{...}{not used}
}
\author{Theresa Albrecht}
\seealso{\code{\link{crossVal}}}
\keyword{methods}
