\name{summary.gpData}
\alias{summary.gpData}
\alias{print.summary.gpData}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary for class gpData
}
\description{
S3 \code{summary} method for objects of class \code{gpData}
}
\usage{
\method{summary}{gpData}(object,...) 
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
object of class \code{gpData}
}
\item{...}{
not used
}
}

\author{
Valentin Wimmer
}

\examples{
\dontrun{
data(maize)
summary(maize)
}
}
\keyword{methods}

