\name{summary.gpData}
\alias{summary.gpData}
\alias{print.summary.gpData}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summarizing object of class gpData
}
\description{
\code{summary} method for class ''gpData''
}
\usage{
\method{summary}{gpData}(object,...) 
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
object of class ''gpData''
}
\item{...}{
not used
}
}

\author{
Valentin Wimmer
}

\examples{
data(maize)
summary(maize)
}
\keyword{methods}

