% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.LD.r
\name{summary.LDdf}
\alias{summary.LDdf}
\alias{summary.LDmat}
\title{Summary for LD objects}
\usage{
\method{summary}{LDdf}(object, cores = 1, ...)
}
\arguments{
\item{object}{object of class \code{LDdf} or \code{LDmat} which is the
output of function \code{pairwiseLD} and argument \code{type="data.frame"}
or \code{type="matrix"}}

\item{cores}{\code{numeric}. Specifies the number of cores for parallel
computing.}

\item{...}{not used}
}
\description{
\code{Summary} method for class ''LDdf'' and ''LDmat''
}
\details{
Returns for each chromosome: Number of markers; mean, minimum and maximum LD
measured as \eqn{r^2}{r2}; fraction of markers with \eqn{r^2>0.2}{r2>0.2};
average and maximum distance of number of markers
}
\examples{

\dontrun{
library(synbreed)
data(maize)
maizeC <- codeGeno(maize)
maizeLD <- pairwiseLD(maizeC, chr = 1:10, type = "data.frame")
maizeLDm <- pairwiseLD(maizeC, chr = 1:10, type = "matrix")
summary(maizeLD)
summary(maizeLDm)
}

}
\seealso{
\code{\link{pairwiseLD}}, ~~~
}
\author{
Valentin Wimmer and Hans-Juergen Auinger
}
