% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synbreed-package.R
\docType{package}
\name{synbreed-package}
\alias{synbreed-package}
\title{synbreed package: A Framework for the Analysis of Genomic
Prediction Data Using R}
\description{
The data flow in synbreed is guided by a single, unified data object of class
\code{gpData} (‘genomic prediction Data’) which is used for storage of multiple data
sources. This includes an array for the phenotypes (individual × trait ×
replication) and a matrix for the marker genotypes (individual × marker
scores). If required, this structure can be extended to include pedigree
information and a marker map. All analysis functions are based on this data
structure. A key feature of the synbreed package is the generality of the
class \code{gpData} which is suitable for a wide range of statistical methods using
genotypic and phenotypic data. Moreover, it is very convenient to store and
share objects of class gpData. Any object of class gpData can be converted to
class \code{cross} in the R package 'qtl' (Broman et al., 2003) and vice versa or to a
\code{data.frame}.
}
\details{
Main Functionality: (draft form)

- Data Preparation
- Visualisation
- \code{create.gpData}
- \code{gp.Mod}
- Genetic Modelling
- and more
}
\examples{
# (better put something here)
}
\keyword{package}
