\name{synbreed-package}
\Rdversion{1.1}
\alias{synbreed-package}
\alias{synbreed}
\docType{package}
\title{
Analysis pipeline for genomic prediction using R
}
\description{
Simulation, analysis and statistivs for genotypic and phenotypic data in plants and animal breeding.
}                                                                                                                                        
\details{
\tabular{ll}{
Package: \tab synbreed\cr                           
Type: \tab Package\cr
Version: \tab 0.3-2\cr
Date: \tab 2010-11-22\cr
License: \tab GPL-2 \cr
LazyLoad: \tab yes\cr
}

}
\author{
Valentin Wimmer
\email{Valentin.Wimmer@wzw.tum.de}

Theresa Albrecht
\email{theresa.albrecht@wzw.tum.de}
}

\keyword{package}


