\name{synbreed-package}
\Rdversion{1.1}
\alias{synbreed-package}
\alias{synbreed}
\docType{package}
\title{
A package containing methods for plant and animal breeding
}
\description{
Simulation, analysis and descriptive funtions for genotypic and phenotypic data in plants and animal breeding.
}                                                                                                                                        
\details{
\tabular{ll}{
Package: \tab synbreed\cr                           
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-03-05\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{
Valentin Wimmer
\email{Valentin.Wimmer@wzw.tum.de}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{

}

