\name{synbreed-package}
\Rdversion{1.1}
\alias{synbreed-package}
\alias{synbreed}
\docType{package}
\title{
A package containing methods for plant and animal breeding
}
\description{
Simulation, analysis and descriptive funtions for genotypic and phenotypic data in plants and animal breeding.
}                                                                                                                                        
\details{
\tabular{ll}{
Package: \tab synbreed\cr                           
Type: \tab Package\cr
Version: \tab 0.3-4\cr
Date: \tab 2010-11-16\cr
License: \tab GPL-2 \cr
LazyLoad: \tab yes\cr
}

}
\author{
Valentin Wimmer
\email{Valentin.Wimmer@wzw.tum.de}
}

\keyword{package}


